+++ src/package/themes/default/index.ts
@@ -4,6 +4,7 @@ import card from './components/card';
 import input from './components/input';
 import passwordInput from './components/password-input';
 import otpInput from './components/otp-input';
+import colorPicker from './components/color-picker';
 import numberInput from './components/number-input';
 import formField from './components/form-field';
 import textarea from './components/textarea';
@@ -57,6 +58,7 @@ export default {
         input,
         passwordInput,
         otpInput,
+        colorPicker,
         numberInput,
         formField,
         textarea,